Move SaltedTxidHasher and SaltedOutPointHasher to saltedhash.{cpp/h}
